refactor: remove Render photo function into SearchResult to make only one call of dispatch
