+++ react-components/src/components/pages/MainPage/components/SearchResult.tsx
@@ -1,11 +1,14 @@
 import React, { useContext } from 'react';
-import { IPhoto, IPhotos, IResponse } from 'types';
+import { DetailType, IPhoto, IPhotos, IResponse } from 'types';
 import isEmpty from 'helpers/isEmpty';
-import RenderPhoto from './RenderPhoto';
 import Context from 'store/Context';
+import PhotoCard from './PhotoCard';
+import getDetailContent from '../functions/getDeteilContent';
 
 function SearchResult(props: {}): JSX.Element {
   const responseObj = useContext(Context).states.photoCardState.responseObj;
+  const dispatch = useContext(Context).dispatches.detailDispatch;
+
   const renderSearchResult = (): JSX.Element[] => {
     let photosArr: IPhoto[] = [];
 
@@ -19,8 +22,22 @@ function SearchResult(props: {}): JSX.Element {
 
     let photoCardsArr: Array<JSX.Element> = [];
     if (photosArr.length > 0) {
-      photoCardsArr = photosArr.map((photo: IPhoto | []): JSX.Element => {
-        return RenderPhoto(photo as IPhoto);
+      photoCardsArr = photosArr.map((photo: IPhoto): JSX.Element => {
+        const srcMedium = `https://live.staticflickr.com/${photo.server}/${photo.id}_${photo.secret}_z.jpg`;
+        const srcLarge = `https://live.staticflickr.com/${photo.server}/${photo.id}_${photo.secret}_c.jpg`;
+        return (
+          <PhotoCard
+            onClick={(event: React.MouseEvent<HTMLDivElement>) => {
+              event.stopPropagation();
+              const detailContent = getDetailContent(photo, srcLarge);
+              dispatch({ type: DetailType.SaveDetailContent, newDetailState: detailContent });
+            }}
+            key={+photo.id}
+            src={srcMedium}
+            title={photo.title}
+            owner={photo.owner}
+          />
+        );
       });
     }
